refactor(assets): decouple asset discovery and download logic
The `src/assets/index.ts` file had grown into a monolithic hub handling
types, resolution strategies, and high-level download orchestration.
This lack of separation caused circular dependencies (particularly
involving `AssetIdentifier` and downstream service imports) and made
the code difficult to test in isolation.

This change splits the asset logic into specialized modules:
- `types.ts`: Core enums and platform utilities.
- `resolvers.ts`: Strategy-based version and URL resolvers.
- `helpers.ts`: Filesystem helpers (cache dir, listAssets, symlink, etc.).
- `downloaders.ts`: CLI tools' download and installation logic.
- `javaAgentDownloader.ts`: Java agent's distinct resolution chain.
- `index.ts`: Pure re-export barrel.

Pulling helpers out of `index.ts` removes the back-edge through the
barrel from `downloaders.ts`, and isolating the Java agent (which still
uses Maven/GitHub-releases/static fallback) lets the CLI flow evolve
independently.

No functional changes were introduced in this commit.
