+++ test/unit/assets/assetService.test.ts
@@ -7,7 +7,8 @@ import { default as chai, expect } from 'chai';
 import { default as chaiFs } from 'chai-fs';
 import { dirname, join } from 'node:path';
 import AssetService from '../../../src/assets/assetService';
-import { AssetIdentifier, listAssets } from '../../../src/assets';
+import { listAssets } from '../../../src/assets';
+import { AssetIdentifier } from '../../../src/assets/types';
 import ResourceVersions from '../../../resources/versions.json';
 import downloadHttpRetry from '../../../src/assets/downloadHttpRetry';
 import { BundledFileDownloadUrlResolver, cacheDir } from '../../../src/assets';
